Turn ClientInterceptorContext into a struct
...and move it into its own file.
