+++ src/csharp/Grpc.Core/Interceptors/Interceptor.cs
@@ -24,47 +24,6 @@ using Grpc.Core.Internal;
 namespace Grpc.Core.Interceptors
 {
     /// <summary>
-    /// Carries along the context associated with intercepted invocations on the client side.
-    /// This is an EXPERIMENTAL API.
-    /// </summary>
-    public class ClientInterceptorContext<TRequest, TResponse>
-        where TRequest : class
-        where TResponse : class
-    {
-        /// <summary>
-        /// Creates a new instance of <see cref="Grpc.Core.Interceptors.ClientInterceptorContext{TRequest, TResponse}" />
-        /// with the specified method, host, and call options.
-        /// </summary>
-        /// <param name="method">A <see cref="Grpc.Core.Method{TRequest, TResponse}"/> object representing the method to be invoked.</param>
-        /// <param name="host">The host to dispatch the current call to.</param>
-        /// <param name="options">A <see cref="Grpc.Core.CallOptions"/> instance containing the call options of the current call.</param>
-
-        public ClientInterceptorContext(Method<TRequest, TResponse> method, string host, CallOptions options)
-        {
-            Method = method;
-            Host = host;
-            Options = options;
-        }
-
-        /// <summary>
-        /// Gets the <see cref="Grpc.Core.Method{TRequest, TResponse}"/> representing
-        /// the method to be invoked.
-        /// </summary>
-        public Method<TRequest, TResponse> Method { get; }
-
-        /// <summary>
-        /// Gets the host that the currect invocation will be dispatched to.
-        /// </summary>
-        public string Host { get; }
-
-        /// <summary>
-        /// Gets the <see cref="Grpc.Core.CallOptions"/> structure representing the
-        /// call options associated with the current invocation.
-        /// </summary>
-        public CallOptions Options { get; }
-    }
-
-    /// <summary>
     /// Serves as the base class for gRPC interceptors.
     /// This is an EXPERIMENTAL API.
     /// </summary>
